backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
yfinance==0.2.65
pandas==2.3.2
numpy==2.3.2
requests==2.32.5
cachetools==5.5.2
pydantic-settings==2.10.1
google-generativeai==0.8.5
